refactor: 패키지 위치 변경 및 분리하여 의존성 방향 정리
- PaymentV1Dto 는 외부 콜백 파라미터의 역할만 수행
- PgClient 요청에 대해서는 별도의 PgApiDto 를 생성하여 사용
