+++ apps/commerce-api/src/main/java/com/loopers/application/payment/PaymentCommand.java
@@ -1,14 +1,14 @@
 package com.loopers.application.payment;
 
-import com.loopers.interfaces.api.payment.TransactionStatus;
+import com.loopers.domain.payment.PaymentStatus;
 
 public class PaymentCommand {
 
     public record Update(
-            TransactionStatus status,
+            PaymentStatus status,
             String transactionKey
     ) {
-        public static Update create(TransactionStatus status, String transactionKey) {
+        public static Update create(PaymentStatus status, String transactionKey) {
             if (status == null || transactionKey == null || transactionKey.isEmpty()) {
                 throw new IllegalArgumentException("상태와 트랜잭션 키는 필수입니다.");
             }
