+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/payment/PgClientAdapter.java
@@ -1,9 +1,7 @@
 package com.loopers.infrastructure.payment;
 
 import com.loopers.domain.payment.CardPaymentClient;
-import com.loopers.domain.payment.PaymentStatus;
-import com.loopers.interfaces.api.payment.PaymentV1Dto;
-import com.loopers.interfaces.api.payment.TransactionStatus;
+import com.loopers.infrastructure.payment.dto.PgApiDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -24,7 +22,7 @@ public class PgClientAdapter implements CardPaymentClient {
 
     @Override
     public Optional<PaymentResponse> requestPayment(PaymentRequest request) {
-        PaymentV1Dto.Request pgRequest = new PaymentV1Dto.Request(
+        PgApiDto.Request pgRequest = new PgApiDto.Request(
                 request.orderId(),
                 request.cardType(),
                 request.cardNo(),
@@ -33,7 +31,7 @@ public class PgClientAdapter implements CardPaymentClient {
         );
 
         try {
-            PaymentV1Dto.TransactionResponse response = pgClient.createPaymentRequest(pgUserId, pgRequest);
+            PgApiDto.TransactionResponse response = pgClient.createPaymentRequest(pgUserId, pgRequest);
 
             if (response.isFail()) {
                 return Optional.empty();
@@ -41,7 +39,7 @@ public class PgClientAdapter implements CardPaymentClient {
 
             return Optional.of(new PaymentResponse(
                     response.data().transactionKey(),
-                    toDomainStatus(response.data().status())
+                    response.data().status().toPaymentStatus()
             ));
         } catch (Exception e) {
             log.error("PG 결제 요청 실패: {}", e.getMessage(), e);
@@ -52,13 +50,13 @@ public class PgClientAdapter implements CardPaymentClient {
     @Override
     public Optional<FindPaymentsResponse> findPaymentsByOrderId(Long orderId) {
         try {
-            PaymentV1Dto.OrderResponse response = pgClient.findPaymentsByOrderId(pgUserId, orderId);
+            PgApiDto.OrderResponse response = pgClient.findPaymentsByOrderId(pgUserId, orderId);
             if (response.isFail() || response.data() == null) {
                 return Optional.empty();
             }
 
             List<TransactionInfo> domainTransactions = response.data().transactions().stream()
-                    .map(tx -> new TransactionInfo(tx.transactionKey(), toDomainStatus(tx.status())))
+                    .map(tx -> new TransactionInfo(tx.transactionKey(), tx.status().toPaymentStatus()))
                     .toList();
 
             return Optional.of(new FindPaymentsResponse(response.data().orderId(), domainTransactions));
@@ -67,13 +65,5 @@ public class PgClientAdapter implements CardPaymentClient {
             log.error("주문 ID로 결제 내역 조회 중 오류 발생. orderId: {}", orderId, e);
             return Optional.empty();
         }
-    }
-
-    private PaymentStatus toDomainStatus(TransactionStatus pgStatus) {
-        return switch (pgStatus) {
-            case SUCCESS -> PaymentStatus.SUCCESS;
-            case FAILED -> PaymentStatus.FAILED;
-            case PENDING -> PaymentStatus.PENDING;
-        };
     }
 }
