--- apps/commerce-api/src/main/java/com/loopers/interfaces/api/payment/TransactionStatus.java
@@ -1,7 +1,0 @@
-package com.loopers.interfaces.api.payment;
-
-public enum TransactionStatus {
-    PENDING,
-    SUCCESS,
-    FAILED
-}
